+++ src/aux-server/aux-web/aux-player/scene/PlayerSimulation3D.ts
@@ -645,7 +645,7 @@ export class PlayerSimulation3D extends Simulation3D {
             ? new Color(userBackgroundColor)
             : undefined;
         await this.simulation.helper.updateBot(userBot, {
-            tags: { 'aux._userContext': this.context },
+            tags: { 'aux._user.context': this.context },
         });
         await this.simulation.helper.updateBot(userBot, {
             tags: { 'aux._user.channel': this.simulation.id },